refactor move to trait
